update README & refactor
