+++ setup.py
@@ -79,21 +79,7 @@ setup(
         ],
         'flask.commands': [
             'references = oarepo_references.cli:references',
-        ],
-        'invenio_i18n.translations': [
-            'messages = oarepo_references',
-        ],
-        # TODO: Edit these entry points to fit your needs.
-        # 'invenio_access.actions': [],
-        # 'invenio_admin.actions': [],
-        # 'invenio_assets.bundles': [],
-        # 'invenio_base.api_apps': [],
-        # 'invenio_base.api_blueprints': [],
-        # 'invenio_base.blueprints': [],
-        # 'invenio_celery.tasks': [],
-        # 'invenio_db.models': [],
-        # 'invenio_pidstore.minters': [],
-        # 'invenio_records.jsonresolver': [],
+        ]
     },
     extras_require=extras_require,
     install_requires=install_requires,
